#61 feat: 리스트조회 및 팔로우 리스트 확인
팔로우 리스트 가져오기 기능 구현

+++ src/main/java/kr/oshino/eataku/list/model/dto/MyListDto.java
@@ -9,11 +9,11 @@ import lombok.*;
 @ToString
 public class MyListDto {
 
-    private int listNo;
+    private Integer listNo;
 
-    private int listShare;
+    private Long listShare;
 
-    private int memberNo;
+    private Long memberNo;
 
     private String listName;
